Move shared file- and tree-item functions to mixin
