+++ source/win-main/file-manager/util/PopoverFileProps.vue
@@ -54,12 +54,12 @@
 </template>
 
 <script>
-import NumberControl from '../../common/vue/form/elements/Number'
-import SelectControl from '../../common/vue/form/elements/Select'
-import { trans } from '../../common/i18n'
-import formatDate from '../../common/util/format-date'
-import formatSize from '../../common/util/format-size'
-import localiseNumber from '../../common/util/localise-number'
+import NumberControl from '../../../common/vue/form/elements/Number'
+import SelectControl from '../../../common/vue/form/elements/Select'
+import { trans } from '../../../common/i18n'
+import formatDate from '../../../common/util/format-date'
+import formatSize from '../../../common/util/format-size'
+import localiseNumber from '../../../common/util/localise-number'
 
 export default {
   name: 'PopoverFileProps',
